Create msp430Music.c
A tiny script that will produce a theme song that I will not mention.  The frequencies are probably different than the actual song.  Please don't hurt me copyright lawyers.  This generally just demonstrates how a speaker can be used to make noise on the msp430.  Credit to Misclik as well.  You will have to compile this into a .elf most likely.